+++ src/config/album_art.rs
@@ -2,12 +2,14 @@ use serde::{Deserialize, Serialize};
 use strum::Display;
 
 use super::Size;
-use crate::shared::terminal::ImageBackend;
+use crate::{mpd::mpd_client::AlbumArtOrder, shared::terminal::ImageBackend};
 
 #[derive(Debug, Default, Serialize, Deserialize, PartialEq, Eq, Clone)]
 pub struct AlbumArtConfigFile {
     #[serde(default)]
     pub method: ImageMethodFile,
+    #[serde(default)]
+    pub order: AlbumArtOrderFile,
     #[serde(default)]
     pub max_size_px: Size,
     #[serde(default = "super::defaults::disabled_album_art_protos")]
@@ -21,6 +23,7 @@ pub struct AlbumArtConfigFile {
 #[derive(Debug, Default, Clone)]
 pub struct AlbumArtConfig {
     pub method: ImageMethod,
+    pub order: AlbumArtOrder,
     pub max_size_px: Size,
     pub disabled_protocols: Vec<String>,
     pub vertical_align: VerticalAlign,
@@ -40,6 +43,15 @@ pub enum HorizontalAlign {
     #[default]
     Center,
     Right,
+}
+
+#[derive(Default, Display, Debug, Serialize, Deserialize, Clone, Copy, PartialEq, Eq)]
+pub enum AlbumArtOrderFile {
+    #[default]
+    EmbeddedFirst,
+    FileFirst,
+    EmbeddedOnly,
+    FileOnly,
 }
 
 #[derive(Default, Display, Debug, Serialize, Deserialize, Clone, Copy, PartialEq, Eq)]
@@ -87,6 +99,12 @@ impl From<AlbumArtConfigFile> for AlbumArtConfig {
         let size = value.max_size_px;
         AlbumArtConfig {
             method: ImageMethod::default(),
+            order: match value.order {
+                AlbumArtOrderFile::EmbeddedFirst => AlbumArtOrder::EmbeddedFirst,
+                AlbumArtOrderFile::FileFirst => AlbumArtOrder::FileFirst,
+                AlbumArtOrderFile::EmbeddedOnly => AlbumArtOrder::EmbeddedOnly,
+                AlbumArtOrderFile::FileOnly => AlbumArtOrder::FileOnly,
+            },
             max_size_px: Size {
                 width: if size.width == 0 { u16::MAX } else { size.width },
                 height: if size.height == 0 { u16::MAX } else { size.height },